requirements.txt
python-dotenv~=1.0.1
transformers~=4.41.2
discord.py~=2.4.0